moved all tasks to core 1
Since I implemented the MQTT and WiFi connection the sensors on the i2c bus were randomly firing. I decided to pin the tasks to core 1 explictly was most tasks where automatically pinned to Core 0, Core 0 is shared with the WiFi and other system tasks. I believe that this is the cause of problems as it's been stable since.